+++ ios/RNSentryReplayBreadcrumbConverter.h
@@ -10,5 +10,7 @@
 
 + (NSString* _Nullable) getTouchPathMessageFrom:(NSArray* _Nullable) path;
 
+- (id<SentryRRWebEvent> _Nullable)convertFrom:(SentryBreadcrumb *_Nonnull) breadcrumb;
+
 @end
 #endif